new file:   .gitignore 	new file:   Gemfile 	modified:   README.md 	renamed:    read_tweets.rb -> scripts/read_tweets.rb
Added mongo to the gemfile, amended README to include a 'bundle install'
command for 'Getting Started'.

Modified read_tweets to insert the tweets into a mongodb collection and
removed parse_tweets. I might need to redo this portion as I am not sure
if the date fields that twitter give and mongo expect are the same but
will find that out in due course.

Moved read_tweets to a script/ directory